requirements-to-freeze.txt
# Flask
Flask

# Flask components for forms/database/bootstrap
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-WTF
Flask-Bootstrap

# External libraries
twilio
phonenumbers
python-dotenv

# Testing
xmlunittest

# Develoment
flake8
black
